Update MF_Custom-Bounce_v1-7.lua
Custom Version B, added the ability to insert your source list directly into the script as a list of strings. Requires formatting like { 'name1', 'name2', 'name3', .... }. Specific details can be found in the code itself.

This is a custom version made by request, and may not be as user friendly, but expands the functionality. I plan to make this change more user friendly, but this was a quick and easy fix to make.